Move to dedicated files
